Lot size is BigDecimal; monitor exposes getSnapshot

+++ src/main/java/com/boris/fundingarbitrage/coinfilter/CoinFilterConfig.java
@@ -1,6 +1,7 @@
 package com.boris.fundingarbitrage.coinfilter;
 
+import java.math.BigDecimal;
+
 public record CoinFilterConfig(
-	double min24hVolumeUsdt,
-	double maxAffordablePrice
+				double min24hVolumeUsdt, BigDecimal maxAffordablePrice
 ) {}
