+++ src/main/java/com/boris/fundingarbitrage/model/exchange/ExchangeSnapshot.java
@@ -5,6 +5,8 @@ import com.boris.fundingarbitrage.model.contract.Fees;
 import com.boris.fundingarbitrage.model.contract.FundingRate;
 import com.boris.fundingarbitrage.model.contract.MarkPrice;
 
+import java.math.BigDecimal;
+
 public record ExchangeSnapshot(
-				BookTicker bookTicker, Fees fees, FundingRate fundingRate, MarkPrice markPrice, double lotSize
+				BookTicker bookTicker, Fees fees, FundingRate fundingRate, MarkPrice markPrice, BigDecimal lotSize
 ) {}
